[BUGFIX] Respect visibility of immediate parent gridelements container
When content elements are located within a disabled gridelements
container, they were not visible in frontend but ke_search still
indexed them.

Now, content elements located in a grid which is disabled are ignored
from indexer. This just works for direct children of gridelements,
because this fix works not recursively.

Resolves: #104